move element-type connection dispatch from model.py to solver.py
_beam_connection_candidate, _plate_connection_candidate,
_CONNECTION_HANDLERS, and _find_connection_handler build JointCandidate
instances from a solver's topology result — they belong next to the
solvers they call, not in model.py. TimberModel.compute_topologies()
now just imports and calls _find_connection_handler.

JointCandidate/PlateJointCandidate imports inside the two builder
functions are local (not top-of-file) to avoid a circular import:
joint.py imports JointTopology from this module, and
compas_timber.elements (needed for the Beam/Plate/Panel handler keys)
itself imports JointTopology from here via PlateFastener, so the
Beam/Plate/Panel import is likewise placed after JointTopology's
definition rather than at the top of the file.
