+++ src/compas_timber/connections/solver.py
@@ -452,3 +452,58 @@ class PlateSolverResult(Data):
             self.distance,
             self.location,
         )
+
+
+# ------------------------------------------------------------------
+# element-type dispatch: builds a JointCandidate for an adjacent pair,
+# based on the pair's element types.
+# ------------------------------------------------------------------
+
+# `compas_timber.elements` imports `JointTopology` from this module (e.g. `PlateFastener`), so this
+# import is placed here (after `JointTopology` is already defined above) rather than with the other
+# top-of-file imports, to avoid a circular import.
+from compas_timber.elements import Beam  # noqa: E402
+from compas_timber.elements import Panel  # noqa: E402
+from compas_timber.elements import Plate  # noqa: E402
+
+
+def _beam_connection_candidate(beam_a, beam_b, max_distance):
+    """Builds a :class:`~compas_timber.connections.JointCandidate` for a pair of adjacent beams, or ``None`` if their topology is unknown."""
+    from compas_timber.connections.joint_candidate import JointCandidate  # local import: avoids a circular import (joint.py imports JointTopology from this module)
+
+    result = ConnectionSolver().find_topology(beam_a, beam_b, max_distance=max_distance)
+    if result.topology == JointTopology.TOPO_UNKNOWN:
+        return None
+    # use the beam order determined by find_topology to keep main, cross relationship
+    return JointCandidate(result.beam_a, result.beam_b, topology=result.topology, distance=result.distance, location=result.location)
+
+
+def _plate_connection_candidate(element_a, element_b, max_distance):
+    """Builds a :class:`~compas_timber.connections.PlateJointCandidate` for a pair of adjacent plates/panels, or ``None`` if their topology is unknown."""
+    from compas_timber.connections.joint_candidate import PlateJointCandidate  # local import: avoids a circular import (joint.py imports JointTopology from this module)
+
+    result = PlateConnectionSolver().find_topology(element_a, element_b, tol=TOL.relative, max_distance=max_distance)
+    if result.topology is JointTopology.TOPO_UNKNOWN:
+        return None
+    kwargs = {"topology": result.topology, "a_segment_index": result.a_segment_index, "distance": result.distance, "location": result.location}
+    if result.topology == JointTopology.TOPO_EDGE_EDGE:
+        kwargs["b_segment_index"] = result.b_segment_index
+    return PlateJointCandidate(result.plate_a, result.plate_b, **kwargs)
+
+
+# Registered as (type_a, type_b) -> handler; order-independent (both (a, b) and (b, a) pairs match).
+# To support a new type combination (e.g. beam-to-plate), add a row here once the corresponding
+# topology-detection geometry exists.
+_CONNECTION_HANDLERS = [
+    ((Beam, Beam), _beam_connection_candidate),
+    ((Plate, Plate), _plate_connection_candidate),
+    ((Panel, Panel), _plate_connection_candidate),
+]
+
+
+def _find_connection_handler(element_a, element_b):
+    """Returns the registered handler for the given pair's element types, or ``None`` if unsupported."""
+    for (type_a, type_b), handler in _CONNECTION_HANDLERS:
+        if (isinstance(element_a, type_a) and isinstance(element_b, type_b)) or (isinstance(element_a, type_b) and isinstance(element_b, type_a)):
+            return handler
+    return None
